requirements.txt
gradio>=4.0.0
tensorflow>=2.13.0
numpy>=1.21.0
pillow>=8.3.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
pandas>=1.3.0